requirements.txt
selenium~=4.4.2
motor~=3.0.0
aredis~=1.1.8
pymongo~=4.2.0
tqdm~=4.64.0
schedule~=1.1.0
Jinja2~=3.1.2
python-telegram-bot~=20.0a4
